+++ fiftyone/utils/coco.py
@@ -241,7 +241,10 @@ class COCODetectionDatasetImporter(
     def __next__(self):
         filename = next(self._iter_filenames)
 
-        image_path = self._image_paths_map[filename]
+        if os.path.isabs(filename):
+            image_path = filename
+        else:
+            image_path = self._image_paths_map[filename]
 
         image_dict = self._image_dicts_map.get(filename, None)
         if image_dict is None: